+++ srcs/MineClearance.UI/Constants.cs
@@ -1,0 +1,17 @@
+namespace MineClearance.UI;
+
+/// <summary>
+/// 常量类
+/// </summary>
+public static class Constants
+{
+    /// <summary>
+    /// 视图模型后缀
+    /// </summary>
+    public const string ViewModelSuffix = "ViewModel";
+
+    /// <summary>
+    /// 视图后缀
+    /// </summary>
+    public const string ViewSuffix = "View";
+}