+++ srcs/MineClearance.UI/Program.cs
@@ -1,3 +1,6 @@
+using Avalonia;
+using Microsoft.Extensions.DependencyInjection;
+using MineClearance.Core;
 using System;
 
 namespace MineClearance.UI;
@@ -13,6 +16,14 @@ file static class Program
     [STAThread]
     private static int Main(string[] args)
     {
-        return 0;
+        App.Services = new ServiceCollection()
+            .AddCore()
+            .BuildServiceProvider();
+
+        return AppBuilder.Configure<App>()
+            .UsePlatformDetect()
+            .WithInterFont()
+            .LogToTrace()
+            .StartWithClassicDesktopLifetime(args);
     }
 }